Refactor BootstrapInitializer
